# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ScholarMatch AI API
  version: 1.0.0
  summary: Shared HTTP contract for the ScholarMatch frontend and backend.
  description: |
    The checked-in contract is the source of truth until the FastAPI routes are implemented.
    All identifiers are UUID strings. All timestamps are RFC 3339/ISO-8601 date-time
    strings in UTC. Authenticated operations use a Supabase access token as a Bearer JWT;
    the backend derives the user identity from its verified `sub` claim.

    Collection endpoints use opaque cursor pagination and return `data` plus `pagination`.
    Errors always use `ErrorEnvelope`, including validation and unexpected errors.

    Operations declaring `Idempotency-Key` require a caller-generated, non-secret key.
    The server scopes it to the authenticated actor, method, and path for at least 24 hours.
    Repeating the same key and request payload returns the original status and body with
    `Idempotency-Replayed: true`. Reusing a key with a different payload, or while the
    original request is still in progress, returns `409 IDEMPOTENCY_CONFLICT`.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [Matches]
      summary: List ranked matches for the authenticated student.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Match page, ordered by descending score with a stable ID tie-breaker.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchPage"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /matches/recalculate:
    post:
      operationId: recalculateMatches
      tags: [Matches]
      summary: Reuse, calculate, or queue deterministic matches for the authenticated student.
      responses:
        "200":
          description: Existing calculations reused or a bounded workload calculated immediately.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecalculationResponse" }
        "202":
          description: An idempotent recalculation job was accepted for a larger workload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecalculationResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /matches/{scholarship_id}:
    get:
      operationId: getMatch
      tags: [Matches]
      summary: Get the authenticated student's match for one scholarship.
      parameters:
        - $ref: "#/components/parameters/ScholarshipId"
      responses:
        "200":
          description: Match found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token; the API verifies signature, issuer, audience, expiry, and subject.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned as `pagination.next_cursor`; clients must not parse it.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ScholarshipId:
      name: scholarship_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Malformed request or cursor.
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
    Unauthorized:
      description: Bearer token is missing or invalid.
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
...
    NotFound:
      description: Resource was not found or is not visible to the actor.
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
    Conflict:
      description: State, duplicate resource, or idempotency conflict.
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
...
    RateLimited:
      description: Request rate exceeded.
      headers:
        Retry-After:
          description: Seconds until the client should retry.
          schema: { type: integer, minimum: 1 }
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
    InternalError:
      description: Unexpected server error with no sensitive implementation details.
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details, request_id]
          properties:
            code:
              type: string
              pattern: "^[A-Z][A-Z0-9_]+$"
              examples: [VALIDATION_ERROR]
            message:
              type: string
              description: Safe, human-readable summary.
            details:
              type: array
              description: Safe structured details; empty when no field-specific detail exists.
              default: []
              items:
                type: object
                additionalProperties: false
                required: [code]
                properties:
                  code: { type: string }
                  field: { type: [string, "null"] }
                  message: { type: [string, "null"] }
            request_id:
              type: string
              format: uuid
...
    ScoreComponent:
      type: object
      additionalProperties: false
      required: [name, formula_version, score, weight, weighted_score, evidence]
      properties:
        name:
          type: string
          enum:
            [
              academic_fit,
              eligibility_fit,
              interests_goals,
              experience,
              readiness_timing,
            ]
        formula_version: { type: string, minLength: 1, maxLength: 100 }
        score: { type: number, minimum: 0, maximum: 1 }
        weight: { type: number, minimum: 0, maximum: 1 }
        weighted_score: { type: number, minimum: 0, maximum: 1 }
        evidence: { type: array, items: { type: string } }
    RuleSource:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        source_url: { type: [string, "null"], format: uri }
        summary: { type: [string, "null"], maxLength: 2000 }
    RuleResult:
      type: object
      additionalProperties: false
      required:
        [
          rule_id,
          field,
          operator,
          strength,
          outcome,
          reason_code,
          message,
          missing_profile_fields,
          source,
          rule_version,
        ]
      properties:
        rule_id: { type: [string, "null"], format: uuid }
        field: { type: string }
        operator: { type: string }
        strength: { type: string, enum: [hard, soft] }
        outcome: { type: string, enum: [eligible, ineligible, unknown] }
        reason_code: { type: string }
        message: { type: string }
        missing_profile_fields: { type: array, items: { type: string } }
        source: { $ref: "#/components/schemas/RuleSource" }
        rule_version: { type: integer, minimum: 1 }
    MatchScholarshipSummary:
      type: object
      additionalProperties: false
      required: [id, title, provider, deadline, funding_type]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        provider: { type: string }
        deadline: { type: [string, "null"], format: date }
        funding_type: { type: string }
    MatchListItem:
      type: object
      additionalProperties: false
      required:
        [
          id,
          scholarship,
          rank_score,
          confidence,
          eligibility,
          missing_profile_fields,
          algorithm_version,
          profile_data_version,
          scholarship_data_version,
          calculated_at,
        ]
      properties:
        id: { type: string, format: uuid }
        scholarship: { $ref: "#/components/schemas/MatchScholarshipSummary" }
        rank_score:
          type: number
          minimum: 0
          maximum: 1
          description: Relative profile-to-scholarship fit score; never a probability of winning.
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence in the score inputs based on profile and requirement completeness; never a probability of winning.
        eligibility: { type: string, enum: [eligible, ineligible, unknown] }
        missing_profile_fields: { type: array, items: { type: string } }
        algorithm_version: { type: string }
        profile_data_version: { type: integer, minimum: 1 }
        scholarship_data_version: { type: integer, minimum: 1 }
        calculated_at: { type: string, format: date-time }
    MatchDetail:
      allOf:
        - $ref: "#/components/schemas/MatchListItem"
        - type: object
          additionalProperties: false
          required: [rule_results, score_breakdown, reasons]
          properties:
            rule_results:
              type: array
              items: { $ref: "#/components/schemas/RuleResult" }
            score_breakdown:
              type: array
              items: { $ref: "#/components/schemas/ScoreComponent" }
            reasons: { type: array, items: { type: string } }
    MatchPage:
      type: object
      additionalProperties: false
      required: [data, next_cursor, has_more, limit]
      properties:
        data:
          { type: array, items: { $ref: "#/components/schemas/MatchListItem" } }
        next_cursor: { type: [string, "null"] }
        has_more: { type: boolean }
        limit: { type: integer, minimum: 1, maximum: 100 }
    RecalculationResponse:
      type: object
      additionalProperties: false
      required:
        [
          mode,
          job_id,
          candidate_count,
          calculated_count,
          reused_count,
          excluded_count,
          profile_data_version,
          algorithm_version,
        ]
      properties:
        mode: { type: string, enum: [existing, immediate, accepted] }
        job_id: { type: [string, "null"], format: uuid }
        candidate_count: { type: integer, minimum: 0 }
        calculated_count: { type: integer, minimum: 0 }
        reused_count: { type: integer, minimum: 0 }
        excluded_count: { type: integer, minimum: 0 }
        profile_data_version: { type: integer, minimum: 1 }
        algorithm_version: { type: string }
...